Rewrites the engagable registry with semaphores

+++ common/src/main/kotlin/registry/EngageableRegistry.kt
@@ -4,7 +4,7 @@ import ai.scynet.common.LifeCycle
 import java.io.Closeable
 
 interface EngageableRegistry<K, V : LifeCycle>: Registry<K, V> {
-    fun engage(key: K): Long
+    fun engage(key: K): Int
 
-    fun disengage(key: K): Long
+    fun disengage(key: K): Int
 }
